Add support for auto_register! syntax
dry-system assumes individual files should require import module
manually, but here we allow this to be skipped, so we must delay
auto_register! behavior until we have Import defined.

This is actually something that we may want to change in dry-system
itself, because it's more inline with how auto-registration works
when you just set dir names in the config.