Make it work with Rails 6.x
This is also a small step back as I had to remove auto-registrar
strategy that works with `app/*` dirs and namespaces corresponding
to the root dirs, so ie `app/operations/foo` having `Operations::Foo`
is no longer supported.

I'm not sure if this will be re-introduced because it turned out this
is against Rails naming conventions so it's probably not worth to have
a feature that tries to work around this convention.