Remove constants from app namespace before rename
When setting up a custom name for the auto inject constant the code left
untouched the previous constant which caused specs and the application
code itself to work when referring to the old constant name. By removing
the constant from app's namespace before redefining it with another name
we solve the problem.